add lsearch in js
