+++ algorithms/ar-binsrt/C++/Bin_Sort.cpp
@@ -1,0 +1,43 @@
+#include <algorithm>
+#include <iostream>
+#include <vector>
+using namespace std;
+ 
+// Function to sort arr[] of
+// size n using bucket sort
+void bucketSort(float arr[], int n)
+{
+     
+    // 1) Create n empty buckets
+    vector<float> b[n];
+ 
+    // 2) Put array elements
+    // in different buckets
+    for (int i = 0; i < n; i++) {
+        int bi = n * arr[i]; // Index in bucket
+        b[bi].push_back(arr[i]);
+    }
+ 
+    // 3) Sort individual buckets
+    for (int i = 0; i < n; i++)
+        sort(b[i].begin(), b[i].end());
+ 
+    // 4) Concatenate all buckets into arr[]
+    int index = 0;
+    for (int i = 0; i < n; i++)
+        for (int j = 0; j < b[i].size(); j++)
+            arr[index++] = b[i][j];
+}
+ 
+int main()
+{
+    float arr[]
+        = {0.8, 0.2, 0.66, 0.5, 0.1, 0.9, 0.336, 0.982, 0.275, 0.9374, 0.125, 0.8754, 0.925};
+    int n = sizeof(arr) / sizeof(arr[0]);
+    bucketSort(arr, n);
+ 
+    cout << "Sorted array is \n";
+    for (int i = 0; i < n; i++)
+        cout << arr[i] << " ";
+    return 0;
+}
